Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
langchain==0.0.340
chromadb==0.4.18
pypdf==3.17.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.1
python-jose==3.3.0
PyMuPDF==1.23.7
pydantic-settings==2.0.3
openai
tiktoken